optimized coding logic
